Metal 2.0 structural: split Compute/DM configs into own headers
ComputeConfiguration and DataMovementConfiguration were defined inline at
the top of kernel_spec.hpp, which had grown unwieldy. Extract each into
its own dedicated header:
  - compute_configuration.hpp
  - data_movement_configuration.hpp

kernel_spec.hpp #includes both; struct contents are byte-identical to
the previous in-place definitions (incl. the disable_implicit_sync_for
field that landed via #45160 on Gen2DataMovementConfig).
